net/http: use copyBufPool in transferWriter.doBodyCopy()
This patch is a followup to 6fd82d83d697edebbe826a028e752c96f195111f. It applies copyBufPool optimization to transferWriter.doBodyCopy(). The function is very frequently used - everytime Request or Response is written to a wire. Without this patch for every Request and Response processed, if there is a body, we need to allocate and GC a 32k buffer. This is quickly causing GC pressure.

Fixes #57202

Change-Id: I4c30e1737726ac8d9937846106efd02effbae300
GitHub-Last-Rev: ba09cda8f3e786656544415f755f27ddc71522d0
GitHub-Pull-Request: golang/go#57205