net/http: fix when writeLoop exited still has request in writech
When there is still a request in writech after
writeLoop exits, the request body will not be
closed and will not be retried.
The request is not sent to the server, so it can
be safely retried without regard to idempotence.

Fixes #49621

Change-Id: Ibd03a3aa0156cf1146af0d28cfd603a6774c61da
GitHub-Last-Rev: 922d323d03354864c241a50241cb5d874c757d93
GitHub-Pull-Request: golang/go#57403