misc/cgo/test: use alternate path for 18146 when coverage enabled
Test test18146 (as written) exec's itself (via argv[0]) a large number
of times (passing a dummy arg to short-circuit test execution) to try
to flush out problems that can crop for repeated fork/exec. This
strategy assumes that a self-exec is a cheap/inexpensive operation,
which is not necessarily the case if we're doing a "-coverpkg=all"
run. Try to avoid long test runtimes and machine oversubscription by
execing a different binary of coverage is enabled.

Change-Id: I105563d001a8654feddbcd172c33a3987aa32006