cmd/link: add debugging flag -showidead
DO NOT SUBMIT
DO NOT REVIEW

Add a debugging flag "-showidead" to the linker that prints out a
summary of symbols that are only referenced from init functions. It
can be interesting to look at these symbols as a way to locate
possibly dead initialized globals (e.g. symbols that are initialized
or scribbled on but then never subsequently used). Output format looks
like:

    Referenced from init only:
    -- package <ppath1> --
    0: <symtyp> <symname> t=<symgotype>
    1: <symtyp> <symname> t=<symgotype>
    ...
    -- package <ppath2> --
    2: <symtyp> <symname> t=<symgotype>
    ...

For example, consider this program:

  package main
  var deadmap = map[string]int{"foo": 1, "bar": 2}
  var sublivemap = map[string]int{"blek": 99}
  var livemap = map[string]int{"floop": sublivemap["blek"]}
  func main() {
  	println("hi mom!", livemap["blarg"])
  }

For the code above, 'deadmap' and 'sublivemap' are referenced only
from within main's init function, but since 'sublivemap' is used
during the initialization of 'livemap', it is in fact live and not
dead (thus things printed by -showidead are potentially but not always
dead).

Change-Id: Ic78c754b1ed9b0f3105c7137afbb67dc4e7a4efa