crypto/elliptic: re-apply some deprecation markers
In the context of #52221, two groups of APIs are getting deprecated:

  - those necessary for ECDH, replaced by crypto/ecdh

  - those too low-level, unsafe, and/or rare to make sense in the stdlib,
    including raw curve operations and custom or generic curves

CL 453256 removed deprecation markers from both, but only the former
need to wait two releases for crypto/ecdh to be widely available.

This CL reintroduces the deprecation markers on the latter APIs.

CL 453256 also removed deprecation markers from the CurveParams generic
implementation, which was deprecated in #34648, and is also not being
replaced, so doesn't need to wait.

Change-Id: Id0e11d7faa3a58a1716ce1ec6e2fff97bab96259