[release-branch.go1.18] runtime: revert "call __fork instead of fork on darwin"
A recent comment on #57263 reports an unexplained crash in a cgo program
that is fixed by reverting the __fork fix. We don't have any viable fix for the
os/exec bug at this point, so give up on a fix for the January point releases.

This reverts CL 459179 (commit 07b6ffb79c90).

Fixes #57689.

Change-Id: I3b81de6bded399f47862325129e86a65c83d8e3b