cmd/compile: experimental global mapdeadcode prototype
DO NOT SUBMIT
DO NOT REVIEW

For debugging/experimentation purposes only. Not yet complete.

This patch hacks the compiler's static initialization machinery
to pick out assignments to global maps (e.g.

   var mymap = map[string]int{"foo":1, "bar":2}

and extract the initialization code for the new map into a separate
init function (which is then called from the main init function via a
weak relocation):

   func init() {
      map.init.0(&mymap)
   }

   func map.init.0(m *map[string]int) {
     *m = map[string]int{"foo":1, "bar":2}
   }

Work that still needs to be done:
- add dummy relocation from map to map.init.N (not sure how this
  should be arranged)
- make both relocations from the main init function (e.g. the
  relo to the mapvar and to the map init func) weak
- change linker to visit init functions after deadcode is finished,
  and for each weak relocation to map.init.N, change the relocation
  to point to runtime.mapinitnoop instead

Change-Id: I1fd6fd6397772be1ebd3eb397caf68ae9a3147e9