cmd/compile,cmd/link: enable deadcode of unreferenced large global maps
This patch changes the compiler's static initialization machinery to
pick out large initialization assignments to global maps (e.g.

   var mymap = map[string]int{"foo":1, "bar":2, ... }

and extract the map init code into a separate init function (which is
then called from the main init function via a weak relocation):

   var mymap map[string]int   // KEEP reloc -> map.init.0

   func init() {
      map.init.0() // weak relocation
   }

   func map.init.0() {
     mymap = map[string]int{"foo":1, "bar":2}
   }

The map init outlining is done selectively (only in the case where the
RHS code exceeds a size limit of 20 IR nodes). [TODO: spend some time
determining whether 20 is the right number here.]

In order to ensure that a given map.init.NNN function is included when
its corresponding map is live, we add dummy R_KEEP relocation from the
map variable to the map init function.

In the linker, once deadcode is complete, we look for calls from an
init function to a map.init.NNN function; if the latter function is
not marked reachable then we rewrite the relocation in question so
that it targets a "noop" routine in the runtime. This seems more
palatable than trying to have the linker modify the init routine text
in some way to erase the call.

On the "pro" side:
- allows large dead maps to be deadcoded, which reduces startup time
  and live heap
- can help with compile time by splitting up giant pkg init funcs into pieces

On the "con" side:
- we have more functions to compile
- cost of map init is slightly higher for live maps due to the need to
  call out to the outlined init routine

Change-Id: I1fd6fd6397772be1ebd3eb397caf68ae9a3147e9