cmd/link/internal/ppc64: generate PC-rel trampolines on P10
These remove the dependence on R2, and are safe to use whenever
PC-relative instructions are supported. Alignment is not an issue
since these symbols are guaranteed to be 16B aligned, and the
single prefixed instruction is placed at the start.

This is a step towards eliminating the need to use and maintain the
TOC pointer in R2 when PCrel is supported.

Change-Id: I1b1a7e16831cfd6732b31f7fad8df2a7c88c8f84