cmd/link,cmd/asm,cmd/compile: avoid TOC on PPC64 in some cases WIP
DO NOT SUBMIT

A TOC pointer isn't strictly necessary when compiling for Power10.

With PC-relative relocation generation, R2 can be ignored entirely
when running in Go code. Thus, it is possible to elide the TOC
regeneration and preservation in the prologue and indirect calls
from Go. It noticably shrinks binary size (e.x 70kB on the compile
 binary).

This is effectively only used for the PIE build mode. There are
still many other changes needed to support module or shared modes.

TODO: It is desirable to find an efficient single place the
linker/asm/compiler can share to ask if the TOC pointer is live.

TODO: What other corner cases is this missing?
Change-Id: Id7901030ec76da5e1367242af7c59970fb640ebd