cmd/internal/obj: use prefix insn in MOV* opcodes for GOPPC64=power10
The assembler backend has full control over relocation generation, that
makes it possible to enable TOC-free code generation without compiler
changes, or changes to go assembly code. This allows the toolchain to
avoid mixing TOC and PC-rel code by exposing the difference to users
when building go. Linked mixed code cgo with the internal linker is not
supported today. This does not change any of this behavior.

Specifically, the assembler is updated to generate PC-rel relocations
when compiling for GOPPC64=power10. This is the first step towards
freeing up the R2/TOC-pointer in future patches. TOC is still maintained
when compiling PIC, but is effectively unused in this configuration.

Likewise, this also covers a few cases where it can be advantageous to
use paddi instead of composite sequences like oris/ori/add to implement
"MOVD <big const>, Rx" or "ADD <big const>, Rx, Ry" using the same
framework which dynamically configures optab.

A final noteworthy difference between prefixed and regular load/store
instruction form is the removal the of 4B alignment restriction on word
and larger operations. This can be used to improve codegen in future
patches.

Updates #44549

Change-Id: If59c216d203c3eed963bfa08855e21771e6ed669