cmd/internal/obj/x86: consistently set AttrNoFrame
amd64 and x86 functions only have the NoFrame attribute
when explicitly marked as such in assembly, but the assembler
promotes some functions to NoFrame using the following heuristics:

- The architecture is x86
- The function is NoSplit and its stack size is explicitly set to 0
- The function is leaf and its stack size is explicitly set to 0

This CL updates the the amd64/x86 assembler to set the NoFrame attribute
in any of the previous cases. Assemblers of all other architecture
already do so.

Change-Id: I48fc76c8685a30754267367c0cee779bd8c87b77