cmd/internal/objabi,runtime: record NoFrame func flag
This CL adds the NoFrame function attribute to the function flag
bits recorded by the assembler so it can be used in latter stages.

In particular, this information will be useful to the linker when
generating Structured Exception Handling (SEH) metadata, which needs
to know if a function uses a frame pointer.

Updates #57302

Change-Id: I5d6571de4394845680090b0f0886f85cd3ecad72